feat(core): fill witnesses and outputsData in generating raw transaction
Fill witnesses and outputsData in generating raw transaction with values provided by users

BREAKING CHANGE: Witnesses field is empty before this commit so users have to place witness values
manually, now it is filled automatically with preset values. With this change, users just need to
replace elements in it but not push them into the array.